+++ tango/math/random/engines/KissCmwc.d
@@ -147,7 +147,7 @@ struct KissCmwc(uint cmwc_r=1024U,ulong cmwc_a=987769338UL){
         foreach (ref val;cmwc_q){
             assert(s[i]=='_',"no separator _ found");
             ++i;
-            uint ate;
+            size_t ate;
             val=cast(uint)Integer.convert(s[i..i+8],16,&ate);
             assert(ate==8,"unexpected read size");
             i+=8;
@@ -155,7 +155,7 @@ struct KissCmwc(uint cmwc_r=1024U,ulong cmwc_a=987769338UL){
         foreach (val;[&cmwc_i,&cmwc_c,&nBytes,&restB,&kiss_x,&kiss_y,&kiss_z,&kiss_c]){
             assert(s[i]=='_',"no separator _ found");
             ++i;
-            uint ate;
+            size_t ate;
             *val=cast(uint)Integer.convert(s[i..i+8],16,&ate);
             assert(ate==8,"unexpected read size");
             i+=8;
