+++ tango/util/log/Log.d
@@ -672,7 +672,7 @@ public class Logger : ILogger
 
         final const(char)[] name ()
         {
-                int i = name_.length;
+                size_t i = name_.length;
                 if (i > 0)
                     --i;
                 return name_[0 .. i];
@@ -1340,7 +1340,7 @@ package struct LogEvent
                 assert (s.length > 0);
                 long ms = time.millis;
 
-                int len = s.length;
+                size_t len = s.length;
                 do {
                    s[--len] = cast(char)(ms % 10 + '0');
                    ms /= 10;
